refactor(settings): fold Service Status into a General card
A dedicated sidebar section for 3 controls (master toggle, poll interval,
badge) was heavier than needed. Move them into a card in the General
section, mirroring the existing Proxy card (toggle + conditional config),
and drop the ServiceStatus sidebar entry and its dedicated view. The
pinnable Service status menu-bar metric is unchanged. Multi-vendor status
is not on the roadmap yet, so there is no need to reserve a section for it.
