+++ Shared/Models/AppSection.swift
@@ -37,7 +37,6 @@ enum SettingsSection: String, CaseIterable {
     case popover
     case agentWatchers
     case notifications
-    case serviceStatus
 
     var labelKey: String {
         switch self {
@@ -48,7 +47,6 @@ enum SettingsSection: String, CaseIterable {
         case .popover:       "sidebar.popover"
         case .agentWatchers: "sidebar.agentWatchers"
         case .notifications: "sidebar.notifications"
-        case .serviceStatus: "sidebar.serviceStatus"
         }
     }
 
@@ -63,7 +61,6 @@ enum SettingsSection: String, CaseIterable {
         case .popover:       "menubar.dock.rectangle"
         case .agentWatchers: "waveform.path.ecg"
         case .notifications: "bell.fill"
-        case .serviceStatus: "dot.radiowaves.left.and.right"
         }
     }
 }
